requirements.txt
pyyaml>=6.0
pytest>=7.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
dspy>=2.0.0
flask>=2.3.0
